+++ src/main/java/net/minecraft/client/gui/Gui.java
@@ -175,9 +175,9 @@ public class Gui
         GlStateManager.enableTexture2D();
     }
 
-    public void drawCenteredString(FontRenderer fontRendererIn, String text, int x, int y, int color)
+    public void drawCenteredString(FontRenderer fontRendererIn, String text, float x, float y, int color)
     {
-        fontRendererIn.drawStringWithShadow(text, (float)(x - fontRendererIn.getStringWidth(text) / 2), (float)y, color);
+        fontRendererIn.drawStringWithShadow(text, x - fontRendererIn.getStringWidth(text) / 2, y, color);
     }
 
     public void drawString(FontRenderer fontRendererIn, String text, int x, int y, int color)
@@ -185,17 +185,17 @@ public class Gui
         fontRendererIn.drawStringWithShadow(text, (float)x, (float)y, color);
     }
 
-    public void drawTexturedModalRect(int x, int y, int textureX, int textureY, int width, int height)
+    public void drawTexturedModalRect(float x, float y, float textureX, float textureY, float width, float height)
     {
         float f = 0.00390625F;
         float f1 = 0.00390625F;
         Tessellator tessellator = Tessellator.getInstance();
         WorldRenderer worldrenderer = tessellator.getWorldRenderer();
         worldrenderer.begin(7, DefaultVertexFormats.POSITION_TEX);
-        worldrenderer.pos(x, y + height, this.zLevel).tex((float)(textureX) * f, (float)(textureY + height) * f1).endVertex();
-        worldrenderer.pos(x + width, y + height, this.zLevel).tex((float)(textureX + width) * f, (float)(textureY + height) * f1).endVertex();
-        worldrenderer.pos(x + width, y, this.zLevel).tex((float)(textureX + width) * f, (float)(textureY) * f1).endVertex();
-        worldrenderer.pos(x, y, this.zLevel).tex((float)(textureX) * f, (float)(textureY) * f1).endVertex();
+        worldrenderer.pos(x, y + height, this.zLevel).tex(textureX * f, (textureY + height) * f1).endVertex();
+        worldrenderer.pos(x + width, y + height, this.zLevel).tex((textureX + width) * f, (textureY + height) * f1).endVertex();
+        worldrenderer.pos(x + width, y, this.zLevel).tex((textureX + width) * f, textureY * f1).endVertex();
+        worldrenderer.pos(x, y, this.zLevel).tex(textureX * f, textureY * f1).endVertex();
         tessellator.draw();
     }
